pyproject.toml
[tool.poetry]
name = "lounger"
version = "0.1.0"
description = "Next generation automated testing framework."
authors = ["虫师 <defnngj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
pytest-playwright = "^0.5.1"
pytest-req = "^0.2.0"
pyyaml = "^6.0.2"
openpyxl = "^3.1.5"
pymysql = "^1.1.1"
pytest-rerunfailures = "^14.0"
pytest-html = "^4.1.1"
click = "^8.1.7"

[tool.poetry.plugins."pytest11"]
lounger = "lounger.plugin"

[tool.poetry.scripts]
lounger = "lounger.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"